fix redirect edge case
